fix:  down arrow 포지션 섹션 별 밑으로 가게 조정함
